backend/requirements.txt
Flask
flask-restful
flask-cors
mysql-connector-python